+++ wp-sitebuilderone.php
@@ -5,7 +5,7 @@
  * GitHub Plugin URI: https://github.com/sitebuilderone/wp-sitebuilderone
  * GitHub Branch: main
  * Description: ACF, LiveCanvas compatible plugin for Local Business websites
- * Version: 0.0.14.4
+ * Version: 0.0.14.5
  * Author: sitebuilderone.com
  * Author URI: https://github.com/sitebuilderone
  * License: GPL-2.0+
@@ -102,10 +102,11 @@ function sbo_add_services_to_glance($items) {
 
 // Include additional files
 include_once plugin_dir_path(__FILE__) . 'includes/header-footer-codes.php';
-include_once plugin_dir_path(__FILE__) . 'includes/dashboard-instructions.php';
+include_once plugin_dir_path(__FILE__) . 'includes/dashboard-guide.php';
 include_once plugin_dir_path(__FILE__) . 'includes/dashboard-social-media.php';
 include_once plugin_dir_path(__FILE__) . 'includes/shortcodes-social.php';
 include_once plugin_dir_path(__FILE__) . 'includes/acf-conditional-shortcodes.php';
+include_once plugin_dir_path(__FILE__) . 'includes/acf-import-export.php';
 include_once plugin_dir_path(__FILE__) . 'includes/dashboard-widget-tasks.php';
 include_once plugin_dir_path(__FILE__) . 'includes/dashboard-shortcodes.php';